+++ custom/plugins/LieferzeitenAdmin/src/Resources/app/administration/src/module/lieferzeiten-settings/page/lieferzeiten-task-assignment-rule-list/index.js
@@ -26,6 +26,10 @@ Component.register('lieferzeiten-task-assignment-rule-list', {
     },
 
     computed: {
+        hasEditAccess() {
+            return this.acl.can('lieferzeiten.editor') || this.acl.can('admin');
+        },
+
         columns() {
             return [
                 { property: 'name', label: 'Name', inlineEdit: 'string', primary: true },
@@ -66,18 +70,30 @@ Component.register('lieferzeiten-task-assignment-rule-list', {
             this.getList();
         },
         onInlineEditSave(item) {
+            if (!this.hasEditAccess) {
+                return Promise.resolve();
+            }
+
             this.isLoading = true;
             return this.repository.save(item, Shopware.Context.api).then(() => {
                 this.getList();
             });
         },
         onDelete(item) {
+            if (!this.hasEditAccess) {
+                return Promise.resolve();
+            }
+
             this.isLoading = true;
             return this.repository.delete(item.id, Shopware.Context.api).then(() => {
                 this.getList();
             });
         },
         onCreate() {
+            if (!this.hasEditAccess) {
+                return Promise.resolve();
+            }
+
             const entity = this.repository.create(Shopware.Context.api);
             entity.name = 'New rule';
             entity.triggerKey = this.triggerOptions[0];
